rename forge proxy to "logic"
will be less confusion with ProxyWorld etc
