--- src/main/scala/hohserg/dimensional/layers/proxy/ServerProxy.scala
@@ -1,5 +1,0 @@
-package hohserg.dimensional.layers.proxy
-
-class ServerProxy extends CommonProxy {
-
-}
