+++ src/main/scala/hohserg/dimensional/layers/sided/ServerLogic.scala
@@ -1,0 +1,5 @@
+package hohserg.dimensional.layers.sided
+
+class ServerLogic extends CommonLogic {
+
+}
